Use vim's :make to run eslint
This allows for automatic navigation of errors within vim using :cn and :cp

For some reason, it appears that vim has to be restarted before running
perhaps becuse the working directory changes (which it shouldn't be doing).
Has to be run from the root of the project because all of the paths that git
outputs are relative to there.